# perf: optimize Eisenstein arithmetic