--- a/field/eisenstein/eisenstein_test.go
+++ b/field/eisenstein/eisenstein_test.go
@@ -89,6 +89,19 @@ func TestEisensteinReceiverIsOperand(t *testing.T) {
 		genE,
 	))
 
+	properties.Property("Having the receiver as operand (mul by conj) should output the same result", prop.ForAll(
+		func(a, b *ComplexNumber) bool {
+			var c, d ComplexNumber
+			d.Set(a)
+			c.MulByConjugate(a, b)
+			a.MulByConjugate(a, b)
+			b.MulByConjugate(&d, b)
+			return a.Equal(b) && a.Equal(&c) && b.Equal(&c)
+		},
+		genE,
+		genE,
+	))
+
 	properties.TestingRun(t, gopter.ConsoleReporter(false))
 }
 
@@ -105,6 +118,17 @@ func TestEisensteinArithmetic(t *testing.T) {
 	properties := gopter.NewProperties(parameters)
 
 	genE := GenComplexNumber(boundSize)
+
+	properties.Property("Mul(Conjugate) & MulByConjugate should output the same result", prop.ForAll(
+		func(a, b *ComplexNumber) bool {
+			var c, d ComplexNumber
+			c.Conjugate(b).Mul(&c, a)
+			d.MulByConjugate(a, b)
+			return c.Equal(&d)
+		},
+		genE,
+		genE,
+	))
 
 	properties.Property("sub & add should leave an element invariant", prop.ForAll(
 		func(a, b *ComplexNumber) bool {
@@ -203,7 +227,7 @@ func TestEisensteinArithmetic(t *testing.T) {
 
 	properties.Property("norm should always be positive", prop.ForAll(
 		func(a *ComplexNumber) bool {
-			return a.Norm().Sign() >= 0
+			return a.Norm(new(big.Int)).Sign() >= 0
 		},
 		genE,
 	))
@@ -242,6 +266,7 @@ func TestEisensteinHalfGCD(t *testing.T) {
 }
 
 func TestEisensteinQuoRem(t *testing.T) {
+
 	t.Parallel()
 	parameters := gopter.DefaultTestParameters()
 	if testing.Short() {
@@ -251,7 +276,19 @@ func TestEisensteinQuoRem(t *testing.T) {
 	}
 
 	properties := gopter.NewProperties(parameters)
+
 	genE := GenComplexNumber(boundSize)
+
+	properties.Property("Quo & QuoRem should output the same result", prop.ForAll(
+		func(a, b *ComplexNumber) bool {
+			var z1, z2, rem ComplexNumber
+			z1.QuoRem(a, b, &rem)
+			z2.Quo(a, b)
+			return z1.Equal(&z2)
+		},
+		genE,
+		genE,
+	))
 
 	properties.Property("QuoRem should be correct", prop.ForAll(
 		func(a, b *ComplexNumber) bool {
@@ -270,11 +307,13 @@ func TestEisensteinQuoRem(t *testing.T) {
 		func(a, b *ComplexNumber) bool {
 			var z, rem ComplexNumber
 			z.QuoRem(a, b, &rem)
-			return rem.Norm().Cmp(b.Norm()) == -1
+			return rem.Norm(new(big.Int)).Cmp(b.Norm(new(big.Int))) == -1
 		},
 		genE,
 		genE,
 	))
+
+	properties.TestingRun(t, gopter.ConsoleReporter(false))
 }
 
 func TestRegressionHalfGCD1483(t *testing.T) {
@@ -283,8 +322,11 @@ func TestRegressionHalfGCD1483(t *testing.T) {
 	a1, _ := new(big.Int).SetString("-303414439467246543595250775667605759171", 10)
 	c0, _ := new(big.Int).SetString("-432420386565659656852420866390673177323", 10)
 	c1, _ := new(big.Int).SetString("238911465918039986966665730306072050094", 10)
-	a := ComplexNumber{A0: a0, A1: a1}
-	c := ComplexNumber{A0: c0, A1: c1}
+	var a, c ComplexNumber
+	a.A0.Set(a0)
+	a.A1.Set(a1)
+	c.A0.Set(c0)
+	c.A1.Set(c1)
 
 	ticker := time.NewTimer(time.Second * 3)
 	doneCh := make(chan struct{})
@@ -318,7 +360,10 @@ func GenComplexNumber(boundSize int64) gopter.Gen {
 		GenNumber(boundSize),
 		GenNumber(boundSize),
 	).Map(func(values []interface{}) *ComplexNumber {
-		return &ComplexNumber{A0: values[0].(*big.Int), A1: values[1].(*big.Int)}
+		var r ComplexNumber
+		r.A0.Set(values[0].(*big.Int))
+		r.A1.Set(values[1].(*big.Int))
+		return &r
 	})
 }
 
@@ -326,15 +371,50 @@ func GenComplexNumber(boundSize int64) gopter.Gen {
 var benchRes [3]*ComplexNumber
 
 func BenchmarkHalfGCD(b *testing.B) {
-	var n, _ = new(big.Int).SetString("100000000000000000000000000000000", 16) // 2^128
-	a0, _ := rand.Int(rand.Reader, n)
-	a1, _ := rand.Int(rand.Reader, n)
-	c0, _ := rand.Int(rand.Reader, n)
-	c1, _ := rand.Int(rand.Reader, n)
-	a := ComplexNumber{A0: a0, A1: a1}
-	c := ComplexNumber{A0: c0, A1: c1}
+	a0, _ := new(big.Int).SetString("121538263010334165887337363056149355411", 10)
+	a1, _ := new(big.Int).SetString("249054933928109647438301795139995905723", 10)
+	c0, _ := new(big.Int).SetString("289494080943284646970981136889290200995", 10)
+	c1, _ := new(big.Int).SetString("289984679903285760251955664955587354062", 10)
+	a := ComplexNumber{A0: *a0, A1: *a1}
+	c := ComplexNumber{A0: *c0, A1: *c1}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		benchRes = HalfGCD(&a, &c)
+	}
+}
+
+func BenchmarkMul(b *testing.B) {
+	a0, _ := new(big.Int).SetString("121538263010334165887337363056149355411", 10)
+	a1, _ := new(big.Int).SetString("249054933928109647438301795139995905723", 10)
+	c0, _ := new(big.Int).SetString("289494080943284646970981136889290200995", 10)
+	c1, _ := new(big.Int).SetString("289984679903285760251955664955587354062", 10)
+	a := ComplexNumber{A0: *a0, A1: *a1}
+	c := ComplexNumber{A0: *c0, A1: *c1}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		benchRes[0].Mul(&a, &c)
+	}
+}
+
+func BenchmarkNorm(b *testing.B) {
+	a0, _ := new(big.Int).SetString("121538263010334165887337363056149355411", 10)
+	a1, _ := new(big.Int).SetString("249054933928109647438301795139995905723", 10)
+	a := ComplexNumber{A0: *a0, A1: *a1}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		a.Norm(&benchRes[0].A0)
+	}
+}
+
+func BenchmarkQuoRem(b *testing.B) {
+	a0, _ := new(big.Int).SetString("121538263010334165887337363056149355411", 10)
+	a1, _ := new(big.Int).SetString("249054933928109647438301795139995905723", 10)
+	c0, _ := new(big.Int).SetString("289494080943284646970981136889290200995", 10)
+	c1, _ := new(big.Int).SetString("289984679903285760251955664955587354062", 10)
+	a := ComplexNumber{A0: *a0, A1: *a1}
+	c := ComplexNumber{A0: *c0, A1: *c1}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		benchRes[0].QuoRem(&a, &c, benchRes[1])
 	}
 }