# Description

Some optimisations for Eisenstein arithmetic. This saves -50% for Half-GCD which we use in hints for emulated scalar multiplications in gnark.

## Type of change

- [x] New feature (non-breaking change which adds functionality)

# How has this been tested?

Existing tests pass + test for new method `MulByConjugate`.

# How has this been benchmarked?

Macbook Air M1:
```
benchmark              old ns/op     new ns/op     delta
BenchmarkHalfGCD-8     74454         36803         -50.57%
BenchmarkMul-8         175           113           -35.17%
BenchmarkNorm-8        109           60.2          -44.93%
BenchmarkQuoRem-8      1034          396           -61.69%
```

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules